requirements.txt
langchain
sqlalchemy
python-dotenv
pydantic
tenacity
langgraph
langchain-google-genai
langchain-community
duckduckgo-search
selenium
webdriver-manager
click>=8.0.0
rich>=10.0.0
beautifulsoup4
IPython